Restore global_filter feature
Split layer.rs into smaller files
Make EtwLayer and EtwFilter more common
